Separate SelectAvatar to other component
